+++ packages/dds/map/src/.mocharc.memory.js
@@ -3,7 +3,7 @@ module.exports = {
     fgrep: "@Benchmark",
     "node-option": ["expose-gc", "gc-global", "unhandled-rejections=strict"], // without leading "--"
     recursive: true,
-    reporter: "node_modules/@fluid-tools/benchmark/dist/MochaMemoryReporter.js",
+    reporter: "node_modules/@fluid-tools/benchmark/dist/MochaMemoryTestReporter.js",
     "reporter-option": ["reportDir=.memoryTestsOutput/"],
     require: ["@fluidframework/mocha-test-setup"],
     spec: ["dist/test/memory/**/*.spec.js", "--perfMode"],
